+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/doc-files/CHANGELOG.md
@@ -17,6 +17,7 @@ exclude: true
 - Added [`GeckoDisplay.windowInsetsChanged`][131.1].
 - Added [`GeckoView.addWindowInsetsListener`][131.2] and [`GeckoView.removeWindowInsetsListener`][131.3] to allow listening WindowInsets changes of Activity's root window with multiple listeners.
 - ⚠️Increased `compileSdkVersion` to 35 (Android 15)
+- ️️⚠️ Removed deprecated [`GeckoSession.ContentDelegate.onProductUrl`][128.5].
 
 [131.1]: {{javadoc_uri}}/GeckoDisplay.html#windowInsetsChanged
 [131.2]: {{javadoc_uri}}/GeckoView.html#addWindowInsetsListener
@@ -1602,4 +1603,4 @@ to allow adding gecko profiler markers.
 [65.24]: {{javadoc_uri}}/CrashReporter.html#sendCrashReport(android.content.Context,android.os.Bundle,java.lang.String)
 [65.25]: {{javadoc_uri}}/GeckoResult.html
 
-[api-version]: b5bac8a734b241d5e3d21f1aaa61ee81f75e32c4
+[api-version]: 057ce230ff9f6b61acad5cd157997ad28f97f637